Finished PassScreen and hooked up transitions
Added functions to change the models on the pass tiers, Added transitions and fixed line up of pass tiers.

Need to add function to let the player level up and unlocking the tiers and link those up to the future locker screen, Task for later.

Tried to add a function to stop quick switch bug when clicking too fast, didn't work. Ask for help