Worked more on settings screen and started Locker screen
Settings screen now has an FPS toggle with an FPS counter and text to go with it. Other display settings with buttons ready to be implemented. Locker screen started with the first panel shown and model ready, need to implement item changing and icons etc. Gonna be a long day.