Finished Shop screen
Shop screen set up and working.
Hooked up buttons so they take away the points, play SFX on both a full purchase AND if the purchase fails for the player not having enough points it plays a Sound effect and shows a panel. Each shop button also sets the corresponding button active in the locker.